Started implementation to LinkedList, Node, Tile and qwirkle files.
Started implementation on the Linked List and various methods that may be needed(such as adding to the front/back of the list).

Node and Tile files have constructors, with Tile also having getter methods.

Qwirkle file only contains a testing method to show that the methods in LinkedList work.